Correcting and Improving the Rectangle Program.
The bug was located in the Point constructor, assigning this.x to y. To
improve the ability to find this bug, I added getters for the point
values assigned to x and y and an accompanying PointTest class with test
cases for those methods. I also added getLength and getWidth methods to
the Rectangle class to simplify the computation along with test cases in
RectangleTest to ensure the proper functioning of those methods.